third_party/WebKit/LayoutTests/http/tests/inspector-protocol/network/interception-take-stream-expected.txt

Tests that resource bodies can be taken as streams.

Running test: testBasicUsage
Error taking stream: Network.takeResponseBodyForInterceptionAsStream is only currently supported with --enable-features=NetworkService

Running test: testLargeRead
Error taking stream: Network.takeResponseBodyForInterceptionAsStream is only currently supported with --enable-features=NetworkService

Running test: testSmallNonOverlappingReads
Error taking stream: Network.takeResponseBodyForInterceptionAsStream is only currently supported with --enable-features=NetworkService

Running test: testSmallOverlappingReads
Error taking stream: Network.takeResponseBodyForInterceptionAsStream is only currently supported with --enable-features=NetworkService

Running test: testRead0
Error taking stream: Network.takeResponseBodyForInterceptionAsStream is only currently supported with --enable-features=NetworkService

Running test: testTakeTwice
Error taking stream: Network.takeResponseBodyForInterceptionAsStream is only currently supported with --enable-features=NetworkService

Running test: testReadAfterCancel
Error taking stream: Network.takeResponseBodyForInterceptionAsStream is only currently supported with --enable-features=NetworkService

Running test: testContinueAfterBodyTaken
Error taking stream: Network.takeResponseBodyForInterceptionAsStream is only currently supported with --enable-features=NetworkService

Running test: testReadAfterClose
Error taking stream: Network.takeResponseBodyForInterceptionAsStream is only currently supported with --enable-features=NetworkService
